test(reactive-rpc): 💍 move folders of E2E tests
